Final Version 1.6 released

+++ linkedList.h
@@ -1,9 +1,8 @@
 #ifndef LINKEDLIST_H
 #define LINKEDLIST_H
 
-
 typedef struct Node{
-    int data;
+    short int data;
     struct Node *next;
 }Node;
